requirements.txt
pandas
Pillow
twine
click
openpyxl